Add code execution license information
